Factored out RunListProgressiveRendering supertype.
